Use free function instead of ObjC methods in attach controller
Summary: Type encodings for all ObjC method signatures are stored as literal strings in the binary. If one of the argument types is something that transitively includes C++ templates, it's type encoding size will baloon very quickly (see P59826536 and [this SO answer](https://stackoverflow.com/questions/11527385/how-are-the-digits-in-objc-method-type-encoding-calculated/11527925#11527925)). Free functions do not have this problem.

Reviewed By: alickbass

Differential Revision: D9126350

fbshipit-source-id: e78c196462e2b375bf78405f7e48a76f00d5dc40