Support component as a target for CKAction
Summary:
Today, when we use scoped components, CKAction will try to execute the code first on the component and then on the controller.

When we use render components, we don't have scopes. Instead, we would just pass the components. It does work in the same way as we traverse the responder chain starting from the target and up in the chain.

The only issue today is that the target actions will assert if the target doesn't have the selector. Hence, we need to update the asserts to support component as a target.

- Added `_CKTypedComponentDebugCheckComponent`
- Refactor `_CKTypedComponentDebugCheckComponentScope` to use the new function
- Refactor `_CKTypedComponentDebugCheckTargetSelector` to support component as a target
- Added tests

Reviewed By: fabiomassimo

Differential Revision: D10217118

fbshipit-source-id: 91f7c2a4595106b04bc67d10b6b76557e5a28452