Fix animations for Render components in the new infra
Summary:
The new animations infra doesn't create component controller for components with animations. However, it relies on `CKComponentScopeHandle` instead and as a result, the animations for render components won't work. The reason is that `CKTreeNode` creates a scope handle only if component has a controller or an initialState.

This diffs makes sure that `CKTreeNode` creates a scope handle in case that a component has animations.
It saves it in a static cache, similar to what we have before in `CKComponent` for the method `controllerClass`.

Reviewed By: kevin0571

Differential Revision: D10345895

fbshipit-source-id: d3645e487c81b7d46d9c4a77ca9ad1d76fc8cc7c