Immediately apply the deferred changeset
Summary: The previous logic was incorrect because it would append the deferred changeset to the existing queue of modifications. The modifications that came after the first part of the changeset but before the second part of the changeset would assume that the whole changeset had been applied and be operating on an invalid data source state. This makes it so that the second part is applied immediately, before any enqueued modifications have the chance to run.

Reviewed By: fabiomassimo

Differential Revision: D10857361

fbshipit-source-id: d65237f301086be87f6c13476770150b156db7c9