Enable update changeset splitting for items out of viewport
Summary: Previously, if all of the items were outside the viewport, it would not defer them at all because it would cause an infinite sequence of deferred changesets being applied. This solution prevents a deferred changesets from being split into further deferred changesets, so splitting for updates out of viewport can be re-enabled.

Reviewed By: ksenks

Differential Revision: D13390971

fbshipit-source-id: 9857bbb693fbdc0ac96f73bb002338f0ef456d03