added a new function in CKYogaBasedComponentProtocol to determine whether layout uses custom baseline property kCKComponentLayoutExtraBaselineKey or not
Summary:
To support older use cases of kCKComponentLayoutExtraBaselineKey which sets custom baseline value for a node, added a new method usesCustomBaseline which return YES if that particular component uses kCKComponentLayoutExtraBaselineKey. Default value is NO

Context:
Stack of diff is to make the behaviour of baseline consistent with web standard where default baseline of node is baseline of first child of node.
Currently in CK baseline function is set on every yoga node and in baseline function we check if custom baseline is being set or not if yes we simply return that value else we return the height of node.
In this stack we are now setting baseline function only when it users of CK set a custom baseline function + to support old use cases of custom baseline value.

Reviewed By: alickbass

Differential Revision: D13782753

fbshipit-source-id: d83dc1e6001aa5286248dfff25f121e502444150