now setting baseline function when usesCustomBaseline is true and added tests for baseline behaviour
Summary:
Added some tests and started using usesCustomBaseline for setting baseline function on yoga node

Context:
Stack of diff is to make the behaviour of baseline consistent with web standard where default baseline of node is baseline of first child of node.
Currently in CK baseline function is set on every yoga node and in baseline function we check if custom baseline is being set or not if yes we simply return that value else we return the height of node.
In this stack we are now setting baseline function only when it users of CK set a custom baseline function + to support old use cases of custom baseline value.

Reviewed By: alickbass

Differential Revision: D13783365

fbshipit-source-id: 740cac0a7f35da24aca7db000ef4b62b0fffcbab