Added equality for bounds animations
Summary:
We have a warning that is triggered when there are multiple bounds animations for a changeset. Based on T40984593, this often fires when these animations are actually equal, which should be completely benign.

This diff defines equality for `CKComponentBoundsAnimation` instances, so we can trigger the warning only if the animations are actually different.

Reviewed By: kevin0571

Differential Revision: D14241808

fbshipit-source-id: 63cb8e8aa5f1a9fce843593679da91f7b4d06bb9