Expose an internal interface for creating a data source item without a layout
Summary:
Currently all data source items have their layout created up front. However
we'd like to be able to change this to decouple the creation of the layout from the
item which is vended to the collection view datasource. This will enable us for
example to tell the collection view datasource about the existence of items prior
to having fully computed their layout.

Reviewed By: kfirapps

Differential Revision: D14572321

fbshipit-source-id: b7d37806c0cd4b57644a703c0680aa1186d49855