Add systrace listener for when we're blocking the main thread on component generation
Summary:
The new parallel layout stuff purposefully makes it where the main thread will
sometimes block on component generation which is taking place on a background thread.
In order to be able to understand how much this is happening and how long the stalls are
we should measure this in an instrumented block. This diff adds the code to enable that.

Reviewed By: kfirapps

Differential Revision: D14577876

fbshipit-source-id: ce6aa64f8441de4f4d2759ea7cf02df49afcd700