Mapping optional to pointer
Summary:
You can't use `map()` with a function returning a pointer since we'll try to construct the resulting optional with a pointer type, which is disallowed.

This diff adds another function — `mapToPtr()` which allows you to transform the wrapped value of an optional but returns the result as a pointer, not as another optional.

Reviewed By: aCorrado

Differential Revision: D15182082

fbshipit-source-id: 8aa16df3ee5e5e6d61c0c52cd448fef602de800e