Animation API
Summary:
### Timing

The timing information that is common for all animations is represented by `Timing` and `SequenceTiming` structs, and all other animation types inherit from these two. The only difference between these two types is that there's no way to set the duration for sequences. Other properties include:

* delay
* timing function

with the corresponding defaults.

Timing structs make use of [CRTP](https://en.wikipedia.org/wiki/Curiously_recurring_template_pattern) in order to return the builder of the correct (derived) type from its methods, otherwise a reference to the base type will be returned and you won't be able to call subtype methods on it [see [Polymorphic Chaining](https://en.wikipedia.org/wiki/Curiously_recurring_template_pattern#Polymorphic_chaining)].
Note that even though there's a default duration of 0.25s for CA animations, this default value is not set in `Timing` but the duration is made explicitly optional instead. This is to avoid setting the duration to animations that are expected to inherit their duration from an animation group they are in.

### Basic Animations

`Basic[Initial|Final|Change]` templates inherit from `Timing` and represent basic animations (in CA sense). Each basic animation type has the associated value type and key path, which:

* makes `from` and `to` values type-safe
* encodes the key path in the type itself

Having the value type as a template parameter allows to have different specialisations for different value types. One example of this would be animations for `CGColor` typed properties that require the ugly casting to `id`. All this can now be hidden from the clients of the framework.

The distinction between initial and final animations also allows us to automatically assign a correct fill mode for the animation: initial animations have the `backwards` fill mode and final animations have the `forwards` fill mode automatically.

Basic animations are separated according to their type in 3 namespaces: `Initial`, `Final` and `Change`. To add support for a new animated property you:

* declare a string constant for the key path
* optionally, add specialisation for the new value type
* declare and implement functions in these namespaces that return an instance of a basic animation with the correct value type and key path

### Combinators

Both `Parallel` and `Sequence` combinators compose two other animations of the same type (initial, change or final). The difference between the two (apart from no way to specify a duration for a sequence) is how the corresponding `CAAnimationGroup` is constructed:

* `Parallel` just puts both animations in a group and sets the `backwards` fill mode if this is an initial animation (just setting the `backwards` fill mode on the individual animations turn out to be insufficient).
* `Sequence` also puts both animations in a group but in addition to that it a) adjusts the begin time of the second animation so it runs after the first and b) sets the duration of the whole group to the sum of individual animation durations. The `backwards` fill mode is also set in case of an initial animation.

Clients of the framework are not supposed to interact with `Parallel` and `Sequence` directly but use factory functions `parallel()` and `sequence()` instead so there's no need to specify the template arguments.

Reviewed By: fabiomassimo

Differential Revision: D15287377

fbshipit-source-id: 5fe7b5d4b06bf08bdd6d478af05a31e93a442ea3