Type-erasing from / to values on construction
Summary:
Reducing code size bloat due to template instantiations. Initial and final builders don't have to be templates too if we immediately convert `from` and `to` values to `id` on construction. The difference in the conversion logic is now handled by an overloaded `animatedValueAsId()` function (this was the only difference between `toCA()` implementations anyway).

On top of that, we don't need to have a whole another specialisation for `UIColor *` for both builders, just the function.

Reviewed By: cuva

Differential Revision: D15495284

fbshipit-source-id: d3766651178d9067ee995ac9e4d53157e1570526