Using fields instead of template args for key paths
Summary:
This is to avoid code size bloat due to template instantiations for every key path. Since animation builders are short-lived and are allocated on the stack, this shouldn't be a problem.

The other observation I had is that `ChangeBuilder` doesn't need value type associated at all, so it's no longer a template.

Reviewed By: alickbass

Differential Revision: D15495213

fbshipit-source-id: 1aa02df59883d82b2254f1df85471a179199ca91