Intercept exception in addGestureRecognizer and provide more info on which component we are mounting
Summary:
T42803971 shows that there is an exception in Apple code when adding a GestureRecognizer in a view while mounting.

This code adds a try/catch block and some logging to find the component that is causing the crash, in order to proceed from there.

The app will still crash

Reviewed By: Andrey-Mishanin

Differential Revision: D16202786

fbshipit-source-id: 3d63be0d8e53be6b44bac4e4441034b9ef3c1cbd