DelayedInitialisationWrapper
Summary:
`CK::DelayedInitialisationWrapper` is an object that allows to store a non-default constructible object as a objective-c ivar. This object expects a value to be set before being read and will assert if this is not the case. This object can also be used to copy assign an object containing const members.

Relies on `CK::Optional` to hold the value.

Reviewed By: kevin0571

Differential Revision: D16620669

fbshipit-source-id: dab539e83bbbf967dca4f69277293dcb0928704a