Use the child's size as the size of the CKCompositeComponent
Summary:
`CKCompositeComponent` cannot have a custom size and the available constructor do not accept custom size. The size that is passed in its constructor is `{}`.

In `CKFlexboxComponent`, however, when we apply the size attributes on the children, we query their size and if a child is `CKCompositeComponent`, it will always have a `{}` size. If deep yoga trees is on, this will result in fact that we will not take into account the size that was set on a child.

In this diff we use the child's `size` as the size of the `CKCompositeComponent` and it fixes the inconsistency

Reviewed By: kevin0571

Differential Revision: D17664640

fbshipit-source-id: 50716ec942a5d43cbaf6b2c58270feb62c9f33a0