Avoid redundant work when applying changeset in datasource events
Summary:
Before this fix, changeset is calculated twice if it's triggered during `didModifyPreviousState` callback.
A flag is added to avoid starting asyncrhonous work in work queue if it's already busy. This causes redundant calculation because the later asynchronous work will just grab a soon to be outdated state and its result will be discarded anyway.

Reviewed By: fabiomassimo

Differential Revision: D18420849

fbshipit-source-id: d75fd3a13d8f2d9113cba716d698c96db0a06d5f