CKVariant: add copy & move ctor/operator=
Summary:
`CK::Variant` doesn't know how to react when being copied / moved since the data is helf as an array `char`. We need to manually specify the copy/move `ctor` and `operator=` and do the logic there.

Added a test which tracks that the retain count of an objective-c object is correct. The intent is to be a proxy to the generated copy/move ctor / operator= to be called.

Reviewed By: aCorrado

Differential Revision: D18921891

fbshipit-source-id: b7c89aa3eb974a4b995c3fc75d1182c3e7d21e00