Adds the ability to match on `CKVariant` with a function pointer
Summary:
`CK::Variant` uses the `Overloard` object to resolve matchers by having a variadic multiple inheritence.

This doesn't work when the type passed is a function pointer. I added `ConditionalFunctionWrapper` which will wrap it in a `std::function` in this case.

Note that we could potentially use another object for that instead of `std::function` since we only need a class and a `operator()` which holds the function pointer.

This is a retry of D18998618 which was reverted because of a missing `#include <functional>`

Reviewed By: fabiomassimo

Differential Revision: D19087335

fbshipit-source-id: b144226b9e7e9302289e7cec340ba09ca6f2aad2