Adding `CKDefines` and excluding all public interface
Summary:
The first step towards swift support is to make `ComponentKit` buildable as a module. To do so, we'll start by excluding the entire public API via these `#if` preprocessors.

Step two will be to enable modules for the target.

Then we'll start by re-including some of the API surface.

Reviewed By: Andrey-Mishanin

Differential Revision: D19558430

fbshipit-source-id: 6583904acf6c18ef7f124b9bc7e65acf446ef84b