Expose CKComponent to Swift.
Summary:
Swift name is now `Component`. Note that IRL it's `ComponentKit.Component` - so that leaves some space for `CKSwift.Component`.

Alternatively we can keep the `CKCompoent` but I couldn't think of a reason to do so.

I've currently excluded `viewContext` and `+newWithView:size:` since we've got no symbols for these yet.

Reviewed By: kfirapps

Differential Revision: D19744076

fbshipit-source-id: 03468eed000cb19468dc98e4d73165cd342757d9