Fix the didReuse logic in CKTreeNode
Summary:
Before:
We were registering the parent from the previous scope root inside the new scope root, even for the new nodes.

After:
We pass explicitly the parent and make sure we register the correct one.

This shouldn't affect correctness as we only need the parent id when traversing up to the root, and the parent id is the same between generations, however, we should definitely fix it.

Reviewed By: Andrey-Mishanin

Differential Revision: D20603148

fbshipit-source-id: a10d101c81acbc2fc3fd53eb34ca4fc175e837d0