Swift wrapper for view configuration
Summary:
Basic support for specifying a view configuration for a component. Only view class and view / layer attributes are supported at the moment.

`ViewConfiguration` is generic with respect to the view type. The type parameter is determined by the view type you pass as the first parameter. This type later determines the root type for all key paths that are used to specify the view attributes. In other words, if `UIView.self` is specified as the view class, you can't use `UIScrollView.contentOffset` attribute. Layer attribute support is currently limited at only specifying `CALayer` key paths.

Another important realisation I had is that the attribute and its value must be unified in a single type, since each key path has an associated value type. That's what `ViewConfiguration.Attribute` and `LayerAttribute` represent.
Moreover, instead of bridging both `CKComponentViewAttribute` and `CKBoxedValue`, only the most general constructor of `CKComponentViewAttribute` taking 3 blocks is bridged. This is enough to build the type-safe keypath-based API.

Reviewed By: aCorrado

Differential Revision: D20838216

fbshipit-source-id: 7b15fb1e75fa035d17dc0ae72e626aa8fb058aed