Fix parent registration for coalesced components
Summary:
I was previously erronously registering the owner as the parent which would lead to weird reuse behaviour.

For context, parent & owners differ when, for example a child component is passed as a prop (like for Flexbox, Inset etc).

Parent ancestors need to be invalidate from the state update as opposed to owner ancestors (although the owner has to be one of the parent ancestors, it might not be the first one).

To keep things simple (and allow run this experiment ASAP) I've re-enabled the second build phase for coalesced composite components.

Let's take things one step at the time. I will be revisiting this later on as an optimisation if worth it - but we need render to nil fixed ASAP.

Reviewed By: kfirapps

Differential Revision: D21383098

fbshipit-source-id: a66937e9528348c33ef4fdcfe5689279b4f9ae16