Properly calculate `parentHasStateUpdate` even if there was no previous component
Summary:
The assumption was that if there was no previous parent this node couldn't possibly have a state update. In reality this can happen if the previous generation didn't contain a render component (thus the second phase wouldn't have been performed and component aren't linked). A state update on a first generation's non reusable component introducing a reusable component in the new generation would lead to an incorrect `parentHasStateUpdate` value.

This wouldn't cause a reuse bug as the current imlementation works in a way that reuse can only happen if the previous tree had their components linked.

Reviewed By: fabiomassimo

Differential Revision: D21617413

fbshipit-source-id: 9a3e341938a857aa2f1219c6c49589001be92fbd