Adding assertions to `CKFlexboxComponent`
Summary:
This is a bit of a shot in the dark. Hopping that if something gets overriden when assertions are enabled we'll be able to have a clearer understanding of the source of the culprits.

My gut feels tells me that this is random memory corruption that just happens because flexbox is omni present in the memory - but I hope I'm wrong.

Reviewed By: kfirapps

Differential Revision: D21548588

fbshipit-source-id: 86782069e2d33ffdac9c5c4f662e550a12ea5508