Use scope responder's unique identifier for id/responder pair in `CKAction`.
Summary:
Much like `CKComponentScopeHandle`, `CKScopedResponder` has a unique identifier.

The unique identifier here is only used to check if an action points to the same instance when deriving an identifier from the action.

While it seems useless to do this change - it will simplify a change in a further diff.

Made this into its own diff to make sure it wasn't lost in other changes.

Reviewed By: kfirapps

Differential Revision: D22136045

fbshipit-source-id: 8c448babd6495f68e822444f1d7709d77f89f1ac