Delay component/controller registration
Summary:
Component registration happens during the first phase of build. Registered components & controllers might thus not be part of the final build component tree (as the component might have been created but then discarded).

In this scenario, the component will be released as not owned by anyone while the controller will still be around (and registered in the scope root for predicates).

Reviewed By: kevin0571

Differential Revision: D23076963

fbshipit-source-id: 2356a1b1ca5ae34aaba64f1b9ac8b52b57ee1967