Remove View generic parameter from `ViewConfiguration`
Summary:
The `View` generic enforces that the view specified as the first argument of `ViewConfiguration.init` matches the key paths supplied in the attributes array. This is a genuine thing which should be enforced but alas bubbling up the generic parameter `View` in the `ViewConfiguration` declaration causes issues when a ViewConfiguration should be returned with no constraint no the view class.

We’d need to have a type erased container to expose a `ViewConfiguration` object from a protocol, for example.

In this case though the object is immutable so only the initialiser requires the constraint to be enforced. Moving the generic type down to the initialiser solves our aforementioned issue (i.e `ViewConfiguration` can be referenced as is without building a type-erased `AnyViewConfiguration`) while still enforcing that the attributes match the view class.

Reviewed By: Andrey-Mishanin

Differential Revision: D23476483

fbshipit-source-id: 4be980a511a2958602dbe92049be2372c6693d4d