Components as Accessibility SoT
Summary:
With this change the accessibility tree is built by using the component hierarchy.
The control from UIKit accessibility is done in CKRootComponentView (that is the root view that contains a component hierarchy) to the mounted component (if any), and it gets back to UIKit in case the component has a mounted view that is also an accessible element or container.

A component subclass can override `accessibilityChildren` to provide a custom ordered list of components that will be used for accessibility (if the component does/should not partecipate to the AX tree, nothing  bad will happen)

The component will also set the correct accessible frame, based on the mounted layout

Reviewed By: ikenwoo

Differential Revision: D24274629

fbshipit-source-id: 61e7fced8613cedbd5c54d94bc8631f010eb34b3