ZStackComponent
Summary:
A layout component that lays out its children on top of each other.

Each child component is sized with the same size constraints as the stack itself and then positioned according to its gravity (see `CKZStackComponentGravity`).

The final size of the stack is a union of all sizes of its children, e.g. if a stack has 2 children with the sizes of {100, 50} and {50, 100}, the resulting stack size would be {100, 100}.

Differential Revision: D24631985

fbshipit-source-id: 09be4a50fbf7389103e9da210ba14374fb55b738